backend/requirements.txt
# EdutechHackathon Backend Dependencies
# Python 3.9+ required

# Core Framework & Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
# sqlite3 is built into Python

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=41.0.0

# AI/ML Integration
openai>=1.6.1
httpx==0.25.2
tiktoken==0.5.2

# File Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
aiofiles==23.2.1
python-magic==0.4.27
Pillow==10.1.0

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
python-dotenv==1.0.0

# Performance & Monitoring
slowapi==0.1.9
structlog==23.2.0

# Optional Production Dependencies
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
sentry-sdk[fastapi]==1.38.0

# Development Environment Helpers
pip-tools==7.3.0
watchfiles==0.21.0  # For uvicorn auto-reload

# Type Stubs for Better Type Checking
types-requests==2.31.0.10
types-redis==4.6.0.11

# Optional Enhanced Features
prometheus-fastapi-instrumentator==6.1.0
langchain==0.0.350
langchain-openai==0.0.2 